[feature] Add implementation to database.py and refactor the test units
