+++ persons/persons.py
@@ -15,9 +15,9 @@ class Person(object):
             id_list.append(person_list[i].ID)
         return id_list
 
-    def generate_id(self, designation, person_list):
+    def generate_id(self, person_list):
         """This function generates id for either staff or fellow."""
-        prefix = "F-" if designation.lower() == "fellow" else "S-"
+        prefix = "F-" if self.designation.lower() == "fellow" else "S-"
         id_exist = True
         person_id = ""
         while id_exist:
